+++ volare/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2022 Efabless Corporation
+# Copyright 2022-2023 Efabless Corporation
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,8 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-__version__ = "0.7.0"
-
-from .common import get_volare_home
+from .common import get_volare_home, get_current_version, get_installed_list
 from .manage import enable
 from .build import build
+
+from .__version__ import __version__
