+++ volare/__version__.py
@@ -11,18 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import List, Dict
+__version__ = "0.7.0"
 
-
-class Family(object):
-    by_name: Dict[str, "Family"] = {}
-
-    def __init__(self, name: str, variants: List[str]):
-        self.name = name
-        self.variants = variants
-
-
-Family.by_name = {}
-Family.by_name["sky130"] = Family("sky130", ["sky130A", "sky130B"])
-Family.by_name["gf180mcu"] = Family("gf180mcu", ["gf180mcuA", "gf180mcuB", "gf180mcuC"])
-Family.by_name["asap7"] = Family("asap7", ["asap7"])
+if __name__ == "__main__":
+    print(__version__, end="")
